+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## 3.3.1 (tba)
 
+### Changes
+
+- Minor under the hood improvements
+
 ### Bug Fixes
 
-- Fixed occassional lag after loading the game, caused by Satsuma rigidbody being enabled when it shouldn't
+- Fixed occassional lag after loading the game, caused by Satsuma rigidbody being enabled when it should not
 - Fixed "crashing" sounds coming from Satsuma and Jonnez when walking next to them
 
 ## 3.3 (20.05.2021)