+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Bug Fixes
 
-- Fixed occassional lag after loading the game, caused by Satsuma rigidbody being enabled when it should not
+- Fixed occassional lag after loading the game, caused by Satsuma rigidbody being enabled when it should not be
 - Fixed "crashing" sounds coming from Satsuma and Jonnez when walking next to them
 
 ## 3.3 (20.05.2021)