+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed "crashing" sounds coming from Satsuma and Jonnez when walking next to them
 - Fixed empty plastic cans possibly being unpickable by the player
 - Fixed MOP disabling empty items, even when "Disable Empty Items" checkbox was not ticked
+- Fixed items despawning too aggressively, if player entered the driving mode and was at home, resulting in items despawing as soon as he entered that mode
 
 ## 3.3 (20.05.2021)